add basic curl implementation testsuite
