--- curl/src/.#tests.rs
@@ -1,1 +1,0 @@
-th0br0@th0lk.10541:1497869525
