+++ curl/src/lib.rs
@@ -3,7 +3,7 @@ extern crate iota_trytes as trytes;
 pub mod constants;
 pub mod curl;
 pub mod simple;
-pub mod tests;
+mod tests;
 
 pub use tests::testsuite;
 pub use curl::*;
