+++ curl/src/simple.rs
@@ -97,66 +97,11 @@ impl Curl for SimpleCurl {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use curl::Curl;
 
     #[test]
-    fn hash_works() {
-        let trans: Trinary = "9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              9999999999999999999999999999999999999999999999999999999999999\
-                              999999999999999999999999999999T999999999999999999999999999999\
-                              99999999999999999999999OLOB99999999999999999999999"
-            .chars()
-            .collect();
-
-        let ex_hash: Trinary = "TAQCQAEBHLLYKAZWMNSXUPWQICMFSKWPEGQBNM9AQMGLFZGME9REOZTQIJQRKYH\
-                             DANIYSMFYPVABX9999"
-            .chars()
-            .collect();
-
-        let mut curl = SimpleCurl::default();
-        curl.absorb(trans);
-        let hash: Trinary = curl.squeeze(HASH_LENGTH);
-
-        assert_eq!(hash, ex_hash);
+    pub fn run_testsuite() {
+        use tests::testsuite;
+        testsuite::run::<SimpleCurl>();
     }
+
 }
