+++ curl/src/tests.rs
@@ -1,13 +1,77 @@
 pub mod testsuite {
-    #[cfg(test)]
-    mod inner {}
+    use curl::*;
 
     #[cfg(test)]
-    pub fn run_testsuite() {
-        use self::inner::*;
+    mod inner {
+        use super::*;
+        use trytes::*;
 
+        pub fn hash_works(curl: &mut Curl) {
+            let trans: Trinary = "9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              9999999999999999999999999999999999999999999999999999999999999\
+                              999999999999999999999999999999T999999999999999999999999999999\
+                              99999999999999999999999OLOB99999999999999999999999"
+                .chars()
+                .collect();
 
+            let ex_hash: Trinary = "TAQCQAEBHLLYKAZWMNSXUPWQICMFSKWPEGQBNM9AQMGLFZGME9REOZTQIJQRKYH\
+                             DANIYSMFYPVABX9999"
+                .chars()
+                .collect();
 
+            curl.absorb(trans);
+            let hash: Trinary = curl.squeeze(HASH_LENGTH);
+
+            assert_eq!(hash, ex_hash);
+        }
+    }
+
+    #[cfg(test)]
+    pub fn run<C: Curl + Default>() {
+        let new_curl = || C::default();
+
+        // run tests
+        inner::hash_works(&mut new_curl());
     }
 
 }
